refactor(admin): change schemas directory structure
