+++ apps/admin/shared/schemas/index.ts
@@ -1,1 +1,1 @@
-export * from './post'
+export * from './CreatePostSchema';
