--- apps/admin/shared/schemas/post/index.ts
@@ -1,1 +1,0 @@
-export * from './CreatePostSchema'
